requirements.txt
Pillow
requests
gpt4all
moviepy
instagrapi
